More gui work for the simulation chamber. Added a pretty dynamic Animation class that will help in chaining animation of strings.

+++ src/main/java/xt9/deepmoblearning/common/items/ItemMobChip.java
@@ -193,7 +193,7 @@ public class ItemMobChip extends ItemBase {
         return ExperienceItem.getCurrentTierKillCountWithSims(getTier(stack), getCurrentTierKillCount(stack), getCurrentTierSimulationCount(stack));
     }
 
-    public static double getCurrentTierSimulationCountWithKills(ItemStack stack) {
+    public static int getCurrentTierSimulationCountWithKills(ItemStack stack) {
         return ExperienceItem.getCurrentTierSimulationCountWithKills(getTier(stack), getCurrentTierKillCount(stack), getCurrentTierSimulationCount(stack));
     }
